(WIP) Emulate the SSH servers on internal nodes
This is needed to enforce a policy on "fleetctl ssh" and "fleetctl
journal". The fleetctl client wants to open a TCP tunnel through the
front-end to reach the internal sshd's, and obviously there is no
letting it do that. Instead, redirect to an ephemeral ssh2.Server
instance that knows about the policy and the steps taken so far
(so that it may figure out exactly where fleetctl intended to land,
instead of trusting the host:port of its TCP forwarding request)

* Refactor authentication logic into an Authenticator helper that can be
  shared between the front-end and internal (simulated) nodes; now
  client has-a Policy only after authentication.
* Also remove sshd.startServer (dead code)